make housekeeping script readable suppress unecessary output from docker-compose

+++ housekeeping/bin/housekeeping.sh
@@ -21,23 +21,46 @@ BULK_DOWNLOAD_EXPIRY="14"     # age in days when matching files should be delete
 WEB_DOWNLOAD_LINK_DIR="${SERVICE_INSTALL_DIR}/monocle_juno_web_root/downloads/"
 WEB_DOWNLOAD_LINK_EXPIRY="30" # age in days when links should be deleted
 
+restart_proxy_container() {
+   local LOG_FILE=/tmp/proxy_restart.out
+   docker-compose restart "$NGINX_CONTAINER" > "$LOG_FILE" 2>&1
+   if [ $? != 0 ]; then
+      echo "Failed to restart container ${NGINX_CONTAINER}:"
+      cat "$LOG_FILE" && rm "$LOG_FILE"
+      exit 255
+   fi
+}
+
+disable_service_access() {
+   mv "$NGINX_SERVICE_CONF" "$NGINX_SERVICE_CONF_TEMP_MOVE"
+   cp "$NGINX_MAINTENACE_CONF" "$NGINX_SERVICE_CONF"
+   restart_proxy_container
+}
+
+enable_service_access() {
+   if [ -f "$NGINX_SERVICE_CONF_TEMP_MOVE" ]; then
+      mv "$NGINX_SERVICE_CONF_TEMP_MOVE" "$NGINX_SERVICE_CONF"
+      restart_proxy_container
+   fi
+}
+
+delete_expired_bulk_download_files() {
+   chmod "$TEMP_READBLE_DIR_MODE" "$BULK_DOWNLOAD_DIR"
+   find "$BULK_DOWNLOAD_DIR" -name "$BULK_DOWNLOAD_GLOB" -follow -mtime +"$BULK_DOWNLOAD_EXPIRY" -exec rm {} \;
+   chmod "$NORMAL_UNREADBLE_DIR_MODE" "$BULK_DOWNLOAD_DIR"
+}
+
+delete_expired_web_download_links() {
+   chmod "$TEMP_READBLE_DIR_MODE" "$WEB_DOWNLOAD_LINK_DIR"
+   find "$WEB_DOWNLOAD_LINK_DIR" -type l -mtime +"$WEB_DOWNLOAD_LINK_EXPIRY" -exec rm {} \;
+   chmod "$NORMAL_UNREADBLE_DIR_MODE" "$WEB_DOWNLOAD_LINK_DIR"
+}
 
 
-# disable access to service at proxy
-mv "$NGINX_SERVICE_CONF" "$NGINX_SERVICE_CONF_TEMP_MOVE"
-cp "$NGINX_MAINTENACE_CONF" "$NGINX_SERVICE_CONF"
-docker-compose restart "$NGINX_CONTAINER"
+disable_service_access
 
-# Delete expired bulk download files
-chmod "$TEMP_READBLE_DIR_MODE" "$BULK_DOWNLOAD_DIR"
-find "$BULK_DOWNLOAD_DIR" -name "$BULK_DOWNLOAD_GLOB" -follow -mtime +"$BULK_DOWNLOAD_EXPIRY" -exec rm {} \;
-chmod "$NORMAL_UNREADBLE_DIR_MODE" "$BULK_DOWNLOAD_DIR"
+delete_expired_bulk_download_files
 
-# Delete expired web download symlinks.
-chmod "$TEMP_READBLE_DIR_MODE" "$WEB_DOWNLOAD_LINK_DIR"
-find "$WEB_DOWNLOAD_LINK_DIR" -type l -mtime +"$WEB_DOWNLOAD_LINK_EXPIRY" -exec rm {} \;
-chmod "$NORMAL_UNREADBLE_DIR_MODE" "$WEB_DOWNLOAD_LINK_DIR"
+delete_expired_web_download_links
 
-# reenable access to service at proxy
-mv "$NGINX_SERVICE_CONF_TEMP_MOVE" "$NGINX_SERVICE_CONF"
-docker-compose restart "$NGINX_CONTAINER"
+enable_service_access
